The __cxa_atexit support is a reqirement for destructor registration of thread-local objects.
For *-*-elf it is already enabled by default.  See comment line 810 in
"gcc/config.gcc".

Define TARGET_LIBGCC_SDATA_SECTION on PowerPC for RTEMS to ".sdata" to
place the __dso_handle.  The __dso_handle is referenced by application
code.  In case this code uses the small data section, the __dso_handle
must be there.

This patch should be committed to GCC 4.8 and 4.9.

Test results:

 href="http://gcc.gnu.org/ml/gcc-testresults/2013-07/msg00671.html