gdb/sparc: SIS gdb output fix, add nouartrx.
Fix the output from the simulator to gdb to use gdb interfaces
so the output is passes through the MI protocol if used.

Add a nouartrx option to the simulator so it does not block on
an stdin read. The termios settings put the tty into non-blocking
mode however the read sometimes blocks on high loads.