Add rtems-bsp-builder.
This is a testing tool that builds BSPs with a range of configure options
for regression testing changes do not break the kernel code.

The builds are controlled by an INI file. The INI configuration has profiles
which define a specific set of architures and BSP to build. There are architectures
which contain BSPs and these further define the options needed to build the BSP.
There is also builds which define the variations each BSP is built with.

The build output can be pointed to any suitable disks so you can control where
the output ends up.

This initial release contains tiers and these are only seeded with something to
test with. It does not define the tiers.

Wanrings, object files and libraries are counted.