bsp-builder: Add support for builds.
Add build support where a build is a combination of options. The
default is 'all' which is a full set of build options passed to
configure. You can now use 'basic' which is the standard or default
configure command line. This used with the arch option lets you
quickly build all BSPs in an architecture.

For example:

 $ rtems-bsp-builder --build-path /builds/rtems/builds/arm \
                     --rtems-tools /opt/rtems/4.12 \
                     --rtems /opt/rtems/src/rtems.git \
                     --arch arm --build basic