rtemstoolkit/libiberty: Fix broken configure detection.
The file libiberty.h requires some configure macros to control some parts
and these are not provided so the file did not match what was needed. Hard
code the result and remove the need for the macros. The related calls
are not provided in our libiberty usage and not needed.

Closes #3103.