Add a target start and target reset to manage boards that do not boot cleanly.
Rename the reset regex as start regex to detect a board has started.
Add a reset regex to detect if a board has failed to boot cleanly and
needs to be reset.

The Microzed I have can fail to boot. The PHY does not initalise.
I do not know if it is the board, the design or u-boot.