bin2c: Add option for alignment
Sometimes it's useful if structures are aligned. This patch add a

  -A alignment

option. Note that this doesn't check for valid alignments. It accepts
any positive number in decimal or hex format. If for example an
alignment of 7 is specified, the compiler will complain that it is not a
power of 2. But it's not really useful to duplicate this check here.