tester/bsps: change stm32h7-stlink to handle SIGTRAP as a nostop
The ST-Link GDB server throws spurious SIGTRAP into the GDB sometimes.
When this happen, the gdb exits immediately as it's run in batch/script
manner. Unfortunately this may be while testcase itself is still running
and does not have enough time to print all the required output.
Such testcase is then marked as failed although otherwise it may run
well to its end.
Adding handle of SIGTRAP as a nostop means that GDB will not exit
after receiving SIGTRAP but rather be forced to continue as nothing
would happen and the running testcase will have a chance to finish
its business.