doc: Add API documentation generation using pdoc
  https://pdoc.dev/

A lot of documentation is missing but it's a start.  If we need more complicated
features in the future we can move to Sphinx but pdoc is great and a very small
dependency.

Remove unused import

This also checks for the existence of a 'python' binary.  It assumes it's Python
2.7 if found and uses it.  If it's not found it skips building the documentation
instead of erroring out.

Once we drop Python 2.7 support we can set this up properly using Wafs internal
Python detection.  It's non-trivial to pickup and support two versions at the
same time.