doxygen: Remove blanket exception
This caused the previous fatal error to not be fatal.  Our doxygen.py was very
out of date and in the latest verison this check is not fatal.

Also fix linkers/wscript to print an error if doxygen is not installed.

This lets all other exceptions pass through and gives us what we want.

Thanks for the help ita.